+++ crates/apps/looper/looper-processor/src/handle/mod.rs
@@ -3,10 +3,9 @@ use std::sync::atomic::{AtomicBool, Ordering};
 use audio_garbage_collector::Handle;
 use audio_processor_traits::{AtomicF32, AudioBuffer};
 use state::LooperProcessorState;
+use state::{LooperQuantizationModeType, RecordingState};
 
-use crate::handle::state::LooperQuantizationModeType;
 use crate::midi_map::MidiMap;
-use crate::RecordingState;
 
 pub mod state;
