+++ crates/apps/looper/looper-processor/src/new_processor/handle/scratch_pad.rs
